refactor: split anthropic provider into module
